bot/requirements.txt
absl-py==2.1.0
adagio==0.2.4
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.11.1
appdirs==1.4.4
astunparse==1.6.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
cloudpickle==3.0.0
contourpy==1.2.1
coreforecast==0.0.8
cycler==0.12.1
Cython==3.0.10
darts==0.29.0
docker==7.0.0
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fs==2.4.16
fsspec==2024.3.1
fugue==0.8.7
fugue-sql-antlr==0.2.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.11.0
holidays==0.47
idna==3.7
Jinja2==3.1.3
joblib==1.4.0
keras==3.2.1
kiwisolver==1.4.5
libclang==18.1.1
lightning-utilities==0.11.2
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
namex==0.0.8
networkx==3.3
nfoursid==1.0.1
numba==0.59.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
ply==3.11
pmdarima==2.0.4
protobuf==4.25.3
pyarrow==15.0.2
Pygments==2.17.2
pyod==1.1.3
Pyomo==6.7.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.2
pytz==2024.1
PyYAML==6.0.1
qpd==0.4.4
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
shap==0.45.0
six==1.16.0
slicer==0.0.7
sqlglot==23.11.2
statsforecast==1.7.4
statsmodels==0.14.2
sympy==1.12
tbats==1.1.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.16.1
# tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.4.0
torch==2.2.2
torchmetrics==1.3.2
tqdm==4.66.2
triad==0.9.6
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
utilsforecast==0.1.5
Werkzeug==3.0.2
wrapt==1.16.0
xarray==2024.3.0
xgboost==2.0.3
yarl==1.9.4
catboost==1.2.5
